requirements.txt
asgiref==3.6.0
Django==4.1.7
django-tables2==2.5.3
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==5.3.0
